+++ backend-rust/src/graphql_api/passive_delegation.rs
@@ -1,34 +1,56 @@
 use super::{
-    baker_and_delegator_types::{DelegationSummary, PaydayPoolReward},
-    get_config, get_pool, ApiResult,
+    baker_and_delegator_types::{CommissionRates, DelegationSummary, PaydayPoolReward},
+    get_config, get_pool, ApiError, ApiResult,
 };
 use crate::{
     connection::{ConnectionQuery, DescendingI64},
-    graphql_api::todo_api,
+    scalar_types::{BigInteger, Decimal},
 };
 use async_graphql::{connection, Context, Object};
+use concordium_rust_sdk::types::AmountFraction;
 use futures::TryStreamExt;
+use sqlx::types::BigDecimal;
 
 #[derive(Default)]
 pub struct QueryPassiveDelegation;
 
 #[Object]
 impl QueryPassiveDelegation {
-    async fn passive_delegation<'a>(&self, _ctx: &Context<'a>) -> ApiResult<PassiveDelegation> {
-        Ok(PassiveDelegation {})
+    async fn passive_delegation<'a>(&self, ctx: &Context<'a>) -> ApiResult<PassiveDelegation> {
+        let pool = get_pool(ctx)?;
+
+        // The `passive_delegation_payday_commission_rates` table has at most one row.
+        // As such taking the `MAX()` of its fields does not change the field
+        // value.
+        let passive_delegation = sqlx::query_as!(
+            PassiveDelegation,
+            "
+                SELECT 
+                    COUNT(*) AS delegator_count,
+                    SUM(delegated_stake) AS delegated_stake,
+                    MAX(payday_transaction_commission) as payday_transaction_commission,
+                    MAX(payday_baking_commission) as payday_baking_commission,             
+                    MAX(payday_finalization_commission) as payday_finalization_commission
+                FROM accounts 
+                    CROSS JOIN passive_delegation_payday_commission_rates
+                WHERE delegated_target_baker_id IS NULL AND
+                    delegated_stake != 0
+            "
+        )
+        .fetch_optional(pool)
+        .await?
+        .ok_or(ApiError::NotFound)?;
+
+        Ok(passive_delegation)
     }
 }
 
-pub struct PassiveDelegation {
-    // commissionRates:  CommissionRates!
-    //
-    // delegatorCount: Int!
-    //
-    // "The total amount staked by delegators to passive delegation."
-    // delegatedStake: UnsignedLong!
-    //
-    // "Total stake passively delegated as a percentage of all CCDs in existence."
-    // delegatedStakePercentage: Decimal!
+struct PassiveDelegation {
+    delegator_count:                Option<i64>,
+    delegated_stake:                Option<BigDecimal>,
+    payday_transaction_commission:  Option<i64>,
+    payday_baking_commission:       Option<i64>,
+    payday_finalization_commission: Option<i64>,
     //
     // Query:
     // apy7days: apy(period: LAST7_DAYS)
@@ -117,47 +139,137 @@ impl PassiveDelegation {
         Ok(connection)
     }
 
-    // Query:
-    // delegators(
-    //     after: $afterDelegators
-    //     before: $beforeDelegators
-    //     first: $firstDelegators
-    //     last: $lastDelegators
-    //   ) {
-    //     nodes {
-    //       accountAddress {
-    //         asString
-    //         __typename
-    //       }
-    //       stakedAmount
-    //       restakeEarnings
-    //       __typename
-    //     }
-    //     pageInfo {
-    //       hasNextPage
-    //       hasPreviousPage
-    //       startCursor
-    //       endCursor
-    //       __typename
-    //     }
-    //     __typename
-    //   }
-    // Schema:
-    // delegators("Returns the first _n_ elements from the list." first: Int
-    // "Returns the elements in the list that come after the specified cursor."
-    // after: String "Returns the last _n_ elements from the list." last: Int
-    // "Returns the elements in the list that come before the specified cursor."
-    // before: String): DelegatorsConnection
+    // Passive delegators are sorted descending by `staked_amount`.
     async fn delegators(
         &self,
-        _ctx: &Context<'_>,
-        #[graphql(desc = "Returns the first _n_ elements from the list.")] _first: Option<u64>,
+        ctx: &Context<'_>,
+        #[graphql(desc = "Returns the first _n_ elements from the list.")] first: Option<u64>,
         #[graphql(desc = "Returns the elements in the list that come after the specified cursor.")]
-        _after: Option<String>,
-        #[graphql(desc = "Returns the last _n_ elements from the list.")] _last: Option<u64>,
+        after: Option<String>,
+        #[graphql(desc = "Returns the last _n_ elements from the list.")] last: Option<u64>,
         #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
-        _before: Option<String>,
-    ) -> ApiResult<connection::Connection<String, DelegationSummary>> {
-        todo_api!()
+        before: Option<String>,
+    ) -> ApiResult<connection::Connection<DescendingI64, DelegationSummary>> {
+        let pool = get_pool(ctx)?;
+        let config = get_config(ctx)?;
+        let query = ConnectionQuery::<DescendingI64>::new(
+            first,
+            after,
+            last,
+            before,
+            config.delegators_connection_limit,
+        )?;
+        let mut row_stream = sqlx::query_as!(
+            DelegationSummary,
+            "SELECT * FROM (
+                SELECT
+                    index,
+                    address as account_address,
+                    delegated_restake_earnings as restake_earnings,
+                    delegated_stake as staked_amount
+                FROM accounts
+                WHERE delegated_target_baker_id IS NULL AND
+                    delegated_stake != 0 AND
+                    accounts.index > $2 AND accounts.index < $1
+                ORDER BY
+                    (CASE WHEN $4 THEN accounts.index END) ASC,
+                    (CASE WHEN NOT $4 THEN accounts.index END) DESC
+                LIMIT $3
+            ) AS delegators
+            ORDER BY delegators.staked_amount DESC",
+            i64::from(query.from),
+            i64::from(query.to),
+            query.limit,
+            query.is_last
+        )
+        .fetch(pool);
+        let mut connection = connection::Connection::new(false, false);
+        while let Some(delegator) = row_stream.try_next().await? {
+            connection.edges.push(connection::Edge::new(delegator.index.into(), delegator));
+        }
+
+        if let (Some(edge_min_index), Some(edge_max_index)) =
+            (connection.edges.last(), connection.edges.first())
+        {
+            let result = sqlx::query!(
+                "
+                SELECT 
+                    MAX(index) as min_index,
+                    MIN(index) as max_index
+                FROM accounts 
+                WHERE delegated_target_baker_id IS NULL
+            "
+            )
+            .fetch_one(pool)
+            .await?;
+
+            connection.has_previous_page =
+                result.max_index.map_or(false, |db_max| db_max > edge_max_index.node.index);
+            connection.has_next_page =
+                result.min_index.map_or(false, |db_min| db_min < edge_min_index.node.index);
+        }
+
+        Ok(connection)
+    }
+
+    async fn delegator_count(&self) -> ApiResult<i64> { Ok(self.delegator_count.unwrap_or(0i64)) }
+
+    async fn delegated_stake(&self) -> ApiResult<BigInteger> {
+        Ok(BigInteger::from(self.delegated_stake.clone().unwrap_or_default()))
+    }
+
+    /// Total passively delegated stake as a percentage of all CCDs in
+    /// existence.
+    async fn delegated_stake_percentage(&self, ctx: &Context<'_>) -> ApiResult<Decimal> {
+        let pool = get_pool(ctx)?;
+
+        let total_ccd_amount: i64 = sqlx::query_scalar!(
+            "
+                SELECT total_amount 
+                FROM blocks 
+                ORDER BY height DESC 
+                LIMIT 1
+            "
+        )
+        .fetch_one(pool)
+        .await?;
+
+        // Division by 0 is not possible because `total_ccd_amount` is always a
+        // positive number.
+        let delegated_stake_percentage: &BigDecimal =
+            &(self.delegated_stake.clone().unwrap_or_default() * 100 / total_ccd_amount);
+
+        let delegated_stake_percentage: Decimal =
+            delegated_stake_percentage.try_into().map_err(|e| {
+                ApiError::InternalError(format!(
+                    "Can not convert `delegated_stake_percentage` to `scalar_types::Decimal`: {}",
+                    e
+                ))
+            })?;
+
+        Ok(delegated_stake_percentage)
+    }
+
+    async fn commission_rates(&self) -> ApiResult<CommissionRates> {
+        let payday_transaction_commission = self
+            .payday_transaction_commission
+            .map(u32::try_from)
+            .transpose()?
+            .map(|c| AmountFraction::new_unchecked(c).into());
+        let payday_baking_commission = self
+            .payday_baking_commission
+            .map(u32::try_from)
+            .transpose()?
+            .map(|c| AmountFraction::new_unchecked(c).into());
+        let payday_finalization_commission = self
+            .payday_finalization_commission
+            .map(u32::try_from)
+            .transpose()?
+            .map(|c| AmountFraction::new_unchecked(c).into());
+        Ok(CommissionRates {
+            transaction_commission:  payday_transaction_commission,
+            baking_commission:       payday_baking_commission,
+            finalization_commission: payday_finalization_commission,
+        })
     }
 }